+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- Update PurchaseButton messages to aid with translation.
 - Update courses and organizations default permissions
 - Update cunningham to 2.0.0
 - Update frontend components OrderStateMessage with new computed state